--- src/Symfony/Component/Security/Core/Authorization/Voter/AccessTrait.php
@@ -1,41 +1,0 @@
-<?php
-
-/*
- * This file is part of the Symfony package.
- *
- * (c) Fabien Potencier <fabien@symfony.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Symfony\Component\Security\Core\Authorization\Voter;
-
-/**
- * @author Dany Maillard <danymaillard93b@gmail.com>
- */
-trait AccessTrait
-{
-    /** @var int One of the VoterInterface::ACCESS_* constants */
-    protected $access;
-
-    public function getAccess(): int
-    {
-        return $this->access;
-    }
-
-    public function isGranted(): bool
-    {
-        return VoterInterface::ACCESS_GRANTED === $this->access;
-    }
-
-    public function isAbstain(): bool
-    {
-        return VoterInterface::ACCESS_ABSTAIN === $this->access;
-    }
-
-    public function isDenied(): bool
-    {
-        return VoterInterface::ACCESS_DENIED === $this->access;
-    }
-}
